Implement ability to insert and reorder lists while retaining element identity.
I'm a little uncertain about this as it introduces a non-O(N) time-complexity for lists with large diffs. In most cases, a list probably won't change all at once, so it'll be ok. But I'd really like a nicer solution. It functions well in practice, though, and should be robust.